Revert "use integer version of pow"
This reverts commit a97666686888ad79e2907101b6d70bc920a6f049.

+++ conversions/octal_to_decimal.c
@@ -2,15 +2,15 @@
 #include <stdio.h>
 
 // Converts octal number to decimal
-int convertValue(int num, int i) { return num * powl(8, i); }
+int convertValue(int num, int i) {
+    return num * pow(8, i);
+}
 
-long long toDecimal(int octal_value)
-{
+long long toDecimal(int octal_value) {
 
     int decimal_value = 0, i = 0;
 
-    while (octal_value)
-    {
+    while (octal_value) {
 
         // Extracts right-most digit and then multiplies by 8^i
         decimal_value += convertValue(octal_value % 10, i++);
@@ -22,8 +22,7 @@ long long toDecimal(int octal_value)
     return decimal_value;
 }
 
-int main()
-{
+int main() {
 
     printf("Enter octal value: ");
